requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
rapidfuzz
pydantic
pillow
pytesseract
python-multipart
streamlit
plotly
requests
